requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
discord.py==1.6.0
dpytest==0.0.22
flake8==3.9.0
idna==3.1
iniconfig==1.1.1
mccabe==0.6.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-cov==2.11.1
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
wavelink==0.9.9
yarl==1.6.3
